+++ tests/functional/weighted/test_weighted_cross_val.py
@@ -2,7 +2,12 @@ import unittest
 
 from scipy.stats import false_discovery_control
 
-from nonconform.estimation.weighted_conformal import WeightedConformalDetector
+from nonconform.estimation.weight import (
+    ForestWeightEstimator,
+    IdentityWeightEstimator,
+    LogisticWeightEstimator,
+)
+from nonconform.estimation.weighted import WeightedConformalDetector
 from nonconform.strategy.cross_val import CrossValidation
 from nonconform.utils.data import Dataset, load
 from nonconform.utils.stat.metrics import false_discovery_rate, statistical_power
@@ -10,12 +15,13 @@ from pyod.models.iforest import IForest
 
 
 class TestCaseSplitConformal(unittest.TestCase):
-    def test_cross_val_conformal_fraud(self):
+    def test_cross_val_conformal_fraud_logistic(self):
         x_train, x_test, y_test = load(Dataset.FRAUD, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
             detector=IForest(behaviour="new"),
             strategy=CrossValidation(k=5, plus=False),
+            weight_estimator=LogisticWeightEstimator(seed=1),
             seed=1,
         )
 
@@ -24,18 +30,19 @@ class TestCaseSplitConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.0, places=1
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.141, places=2
         )
         self.assertAlmostEqual(
-            statistical_power(y=y_test, y_hat=decisions), 0.22, places=2
+            statistical_power(y=y_test, y_hat=decisions), 0.79, places=2
         )
 
-    def test_cross_val_conformal_plus_fraud(self):
+    def test_cross_val_conformal_plus_fraud_forest(self):
         x_train, x_test, y_test = load(Dataset.FRAUD, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
             detector=IForest(behaviour="new"),
             strategy=CrossValidation(k=5, plus=True),
+            weight_estimator=ForestWeightEstimator(seed=1),
             seed=1,
         )
 
@@ -44,18 +51,19 @@ class TestCaseSplitConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.0, places=1
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.153, places=2
         )
         self.assertAlmostEqual(
-            statistical_power(y=y_test, y_hat=decisions), 0.24, places=2
+            statistical_power(y=y_test, y_hat=decisions), 0.83, places=2
         )
 
-    def test_cross_val_conformal_shuttle(self):
+    def test_cross_val_conformal_shuttle_identity(self):
         x_train, x_test, y_test = load(Dataset.SHUTTLE, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
             detector=IForest(behaviour="new"),
             strategy=CrossValidation(k=5, plus=False),
+            weight_estimator=IdentityWeightEstimator(),
             seed=1,
         )
 
@@ -64,18 +72,19 @@ class TestCaseSplitConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.109, places=3
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.168, places=3
         )
         self.assertAlmostEqual(
-            statistical_power(y=y_test, y_hat=decisions), 0.98, places=2
+            statistical_power(y=y_test, y_hat=decisions), 0.99, places=2
         )
 
-    def test_cross_val_conformal_plus_shuttle(self):
+    def test_cross_val_conformal_plus_shuttle_logistic(self):
         x_train, x_test, y_test = load(Dataset.SHUTTLE, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
             detector=IForest(behaviour="new"),
             strategy=CrossValidation(k=5, plus=True),
+            weight_estimator=LogisticWeightEstimator(seed=1),
             seed=1,
         )
 
@@ -84,18 +93,19 @@ class TestCaseSplitConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.093, places=3
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.116, places=2
         )
         self.assertAlmostEqual(
-            statistical_power(y=y_test, y_hat=decisions), 0.98, places=2
+            statistical_power(y=y_test, y_hat=decisions), 0.99, places=2
         )
 
-    def test_cross_val_conformal_thyroid(self):
+    def test_cross_val_conformal_thyroid_forest(self):
         x_train, x_test, y_test = load(Dataset.THYROID, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
             detector=IForest(behaviour="new"),
             strategy=CrossValidation(k=5, plus=False),
+            weight_estimator=ForestWeightEstimator(seed=1),
             seed=1,
         )
 
@@ -104,10 +114,10 @@ class TestCaseSplitConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.067, places=3
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.229, places=2
         )
         self.assertAlmostEqual(
-            statistical_power(y=y_test, y_hat=decisions), 0.459, places=3
+            statistical_power(y=y_test, y_hat=decisions), 0.934, places=2
         )