+++ tests/functional/weighted/test_weighted_randomized.py
@@ -2,7 +2,12 @@ import unittest
 
 from scipy.stats import false_discovery_control
 
-from nonconform.estimation.weighted_conformal import WeightedConformalDetector
+from nonconform.estimation.weight import (
+    ForestWeightEstimator,
+    IdentityWeightEstimator,
+    LogisticWeightEstimator,
+)
+from nonconform.estimation.weighted import WeightedConformalDetector
 from nonconform.strategy.experimental.randomized import Randomized
 from nonconform.utils.data import Dataset, load
 from nonconform.utils.stat.metrics import false_discovery_rate, statistical_power
@@ -12,12 +17,13 @@ from pyod.models.iforest import IForest
 
 
 class TestCaseRandomizedConformal(unittest.TestCase):
-    def test_randomized_conformal_fraud(self):
+    def test_randomized_conformal_fraud_logistic(self):
         x_train, x_test, y_test = load(Dataset.FRAUD, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
             detector=IForest(behaviour="new"),
             strategy=Randomized(n_calib=100_000, plus=True),
+            weight_estimator=LogisticWeightEstimator(seed=1),
             seed=1,
         )
 
@@ -26,18 +32,19 @@ class TestCaseRandomizedConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.0, places=1
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.126, places=3
         )
         self.assertAlmostEqual(
-            statistical_power(y=y_test, y_hat=decisions), 0.08, places=2
+            statistical_power(y=y_test, y_hat=decisions), 0.76, places=2
         )
 
-    def test_randomized_conformal_shuttle(self):
+    def test_randomized_conformal_shuttle_forest(self):
         x_train, x_test, y_test = load(Dataset.SHUTTLE, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
             detector=IForest(behaviour="new"),
             strategy=Randomized(n_calib=100_000, plus=True),
+            weight_estimator=ForestWeightEstimator(seed=1),
             seed=1,
         )
 
@@ -46,18 +53,19 @@ class TestCaseRandomizedConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.101, places=3
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.182, places=3
         )
         self.assertAlmostEqual(
-            statistical_power(y=y_test, y_hat=decisions), 0.98, places=2
+            statistical_power(y=y_test, y_hat=decisions), 0.99, places=2
         )
 
-    def test_randomized_conformal_thyroid(self):
+    def test_randomized_conformal_thyroid_identity(self):
         x_train, x_test, y_test = load(Dataset.THYROID, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
             detector=IForest(behaviour="new"),
             strategy=Randomized(n_calib=10_000, plus=True),
+            weight_estimator=IdentityWeightEstimator(),
             seed=1,
         )
 
@@ -66,18 +74,19 @@ class TestCaseRandomizedConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.057, places=3
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.138, places=3
         )
         self.assertAlmostEqual(
-            statistical_power(y=y_test, y_hat=decisions), 0.82, places=2
+            statistical_power(y=y_test, y_hat=decisions), 0.918, places=3
         )
 
-    def test_randomized_conformal_mammography(self):
+    def test_randomized_conformal_mammography_logistic(self):
         x_train, x_test, y_test = load(Dataset.MAMMOGRAPHY, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
             detector=ECOD(),
             strategy=Randomized(n_calib=100_000, plus=True),
+            weight_estimator=LogisticWeightEstimator(seed=1),
             seed=1,
         )
 
@@ -86,18 +95,19 @@ class TestCaseRandomizedConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.077, places=3
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.065, places=3
         )
         self.assertAlmostEqual(
-            statistical_power(y=y_test, y_hat=decisions), 0.12, places=2
+            statistical_power(y=y_test, y_hat=decisions), 0.29, places=2
         )
 
-    def test_randomized_conformal_musk(self):
+    def test_randomized_conformal_musk_forest(self):
         x_train, x_test, y_test = load(Dataset.MUSK, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
             detector=HBOS(),
             strategy=Randomized(n_calib=10_000, plus=True),
+            weight_estimator=ForestWeightEstimator(seed=1),
             seed=1,
         )
 
@@ -106,7 +116,7 @@ class TestCaseRandomizedConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.155, places=3
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.183, places=3
         )
         self.assertAlmostEqual(
             statistical_power(y=y_test, y_hat=decisions), 1.0, places=1