+++ tests/functional/weighted/test_weighted_split.py
@@ -2,7 +2,11 @@ import unittest
 
 from scipy.stats import false_discovery_control
 
-from nonconform.estimation.weighted_conformal import WeightedConformalDetector
+from nonconform.estimation.weight import (
+    ForestWeightEstimator,
+    IdentityWeightEstimator,
+)
+from nonconform.estimation.weighted import WeightedConformalDetector
 from nonconform.strategy.split import Split
 from nonconform.utils.data import Dataset, load
 from nonconform.utils.stat.metrics import false_discovery_rate, statistical_power
@@ -10,11 +14,14 @@ from pyod.models.iforest import IForest
 
 
 class TestCaseSplitConformal(unittest.TestCase):
-    def test_split_conformal_fraud(self):
+    def test_split_conformal_fraud_forest(self):
         x_train, x_test, y_test = load(Dataset.FRAUD, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
-            detector=IForest(behaviour="new"), strategy=Split(n_calib=10_000), seed=1
+            detector=IForest(behaviour="new"),
+            strategy=Split(n_calib=10_000),
+            weight_estimator=ForestWeightEstimator(seed=1),
+            seed=1,
         )
 
         ce.fit(x_train)
@@ -22,17 +29,20 @@ class TestCaseSplitConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.134, places=3
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.129, places=3
         )
         self.assertAlmostEqual(
-            statistical_power(y=y_test, y_hat=decisions), 0.71, places=2
+            statistical_power(y=y_test, y_hat=decisions), 0.81, places=2
         )
 
-    def test_split_conformal_shuttle(self):
+    def test_split_conformal_shuttle_identity(self):
         x_train, x_test, y_test = load(Dataset.SHUTTLE, setup=True, seed=1)
 
         ce = WeightedConformalDetector(
-            detector=IForest(behaviour="new"), strategy=Split(n_calib=10_000), seed=1
+            detector=IForest(behaviour="new"),
+            strategy=Split(n_calib=10_000),
+            weight_estimator=IdentityWeightEstimator(),
+            seed=1,
         )
 
         ce.fit(x_train)
@@ -40,10 +50,10 @@ class TestCaseSplitConformal(unittest.TestCase):
 
         decisions = false_discovery_control(est, method="bh") <= 0.2
         self.assertAlmostEqual(
-            false_discovery_rate(y=y_test, y_hat=decisions), 0.075, places=3
+            false_discovery_rate(y=y_test, y_hat=decisions), 0.201, places=2
         )
         self.assertAlmostEqual(
-            statistical_power(y=y_test, y_hat=decisions), 0.98, places=2
+            statistical_power(y=y_test, y_hat=decisions), 0.99, places=2
         )